refactored model_read to io/model
