+++ tardis/model/base.py
@@ -10,7 +10,7 @@ from tardis.model.geometry.radial1d import Radial1DGeometry
 
 from tardis.util.base import quantity_linspace, is_valid_nuclide_or_elem
 from tardis.io.parsers.csvy import load_csvy
-from tardis.io.model_reader import (
+from tardis.io.model.model_reader import (
     read_density_file,
     read_abundances_file,
     read_uniform_abundances,
