+++ tardis/visualization/widgets/custom_abundance.py
@@ -20,7 +20,7 @@ from tardis.model.density import (
 )
 from tardis.io.config_validator import validate_dict
 from tardis.io.parsers.csvy import load_csvy
-from tardis.io.model_reader import (
+from tardis.io.model.model_reader import (
     read_uniform_abundances,
     parse_csv_abundances,
 )
